third_party/blink/web_tests/accessibility/contenteditable-notifications.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../resources/run-after-layout-and-paint.js"></script>

<div id="textbox" contenteditable="true">
  First<p>Second</p>
</div>

<script>
'use strict';

async_test((t) => {
  const textbox = document.getElementById('textbox');
  const axTextBox = accessibilityController.accessibleElementById('textbox');
  let valueChangedCount = 0;
  let selectedTextChangedCount = 0;
  const expectedIntents = [];

  axTextBox.setNotificationListener(t.step_func((notification, intents) => {
    if (notification == "ValueChanged") {
      ++valueChangedCount;
    } else if (notification == "SelectedTextChanged") {
      assert_array_equals(intents, [expectedIntents[selectedTextChangedCount]]);
      ++selectedTextChangedCount;
    }

    if (valueChangedCount >= 4 && selectedTextChangedCount >= 9) {
      t.done();
    }
  }));

  runAfterLayoutAndPaint(t.step_func(() => {
    textbox.focus();
    // Initial setting of the selection.
    expectedIntents.push('AXEventIntent(setSelection,character,forward)');
    eventSender.keyDown("ArrowDown", []);
    expectedIntents.push('AXEventIntent(moveSelection,lineStart,forward)');
    eventSender.keyDown("ArrowDown", []);
    expectedIntents.push('AXEventIntent(moveSelection,lineStart,forward)');
    eventSender.keyDown("ArrowLeft", []);
    expectedIntents.push('AXEventIntent(moveSelection,character,backward)');
    eventSender.keyDown("ArrowLeft", []);
    expectedIntents.push('AXEventIntent(moveSelection,character,backward)');
    eventSender.keyDown("w", []);
    // TODO(nektar): Support an intent for the typing command.
    expectedIntents.push('AXEventIntent(setSelection,character,forward)');
    eventSender.keyDown("x", []);
    expectedIntents.push('AXEventIntent(setSelection,character,forward)');
    eventSender.keyDown("y", []);
    expectedIntents.push('AXEventIntent(setSelection,character,forward)');
    eventSender.keyDown("z", []);
    expectedIntents.push('AXEventIntent(setSelection,character,forward)');
  }));
}, 'Moving the cursor in a contentEditable sends a selected text change notification, and typing in a contentEditable sends both a value changed and selected text changed notification.');

</script>
